Update bunny GEM to version 1.5.1
Various fixes have happened since version 1.3.1, but the primary
motivation for upgrading is to get some improvements to reliability of
reconnection to the Rabbit cluster on failure.  In particular, this
allows a list of hosts to be passed to Bunny.new(), instead of a single
one, removing a single point of failure.