Create the queue durably
If not specified, bunny defaults to creating queues non-durably.

Durable queues continue to exist after a broker restart.  Without this,
after a rabbit cluster restart, any messages sent to the exchange before
the alert service was restarted (causing it to re-create the queue)
would be lost.

This also showed up that messages weren't being discarded from the queue
in one of the tests because stubbing was preventing the original
"discard" call happening.  Fixed this.