Inject an identifier into email body
Story: https://www.pivotaltracker.com/story/show/84134884

In order to able to compare sent and received emails via the Email alert monitor, we need
to inject an identifier into the email body.

This PR reuses the implementation for generating an identifier for a
message in Redis, injecting it into the wrapper div in the email body
with a `data-message-id` attribute. The use of an HTML data-attribute tries to
make it easier to retrieve the identifier in other apps, such as the
Email alert monitor.