Handle messages with no title or description
This is in response to errors we've started getting with processing of
redirect documents (which have neither title nor description), and
organisation documents (which have no description).  These document
formats didn't exist in content-store when the service was initially
designed.  Further, until some recent changes in the publishing API,
even when the fields were absent from the document sent to the
publishing API, the message on the message queue had the fields present
with their values set to `nil`, so passed validation here.

"Content" documents should always have a title, so ignoring documents
without them is safe, and has the small benefit of ensuring that there's
no risk of things like "redirect" documents triggering alerts.  Sending
an alert for a document without a title would be uninformative anyway,
since the title is how we identify the document in the body of the
message.

If a document has a missing description field, we can still send a
useful message - it will just be missing the description of the
document.  Therefore, we allow messages to be sent in this case.

The code this touches could do with some refactoring, but I don't want
to get people distracted by that right now - I just want to fix the
immediate problems.