Don't error on redirect items.
Previously, this would error for items without a public_updated_at field
set (ie redirect or gone items).  This is because the check for the
presence of a title happens after the validation.

This has involved some refactoring of the message processor so that it
checks for the title and language before validating the document.