Handle messages which have no details hash
Such messages can be valid (for example, the special_route documents are
valid items).

Rather than raising a validation error for such documents, just ignore
them.

Currently, such documents can never be tagged to topics (since
the topic tags are held in the details hash), but in future it will be
possible for them to be tagged using the links hash, so I've not
implemented this by actively ignoring the documents - instead, I've just
handled a missing details hash the same as one which is present but
empty.