Include correct URL in policy change notification mail
When a policy is changed, an email is sent to everyone who is subscribed to those notifications.
Currently, the url included in this mail points to the sign up for notifications page, and not to the policy itself.
This change will correct this error.

NOTE: This change is temporary - it is achieved by stripping off the "/email-signup" from the end of the URL.  In future,
once the links attribute has been included in the message document, we can interrogate the content store to get the correct link.