# This is a combination of 2 commits. # The first commit's message is:
# This is a combination of 3 commits.
# The first commit's message is:
Include correct URL in policy change notification mail

When a policy is changed, an email is sent to everyone who is subscribed to those notifications.
Currently, the url included in this mail points to the sign up for notifications page, and not to the policy itself.
This change will correct this error.

NOTE: This change is temporary - it is achieved by stripping off the "/email-signup" from the end of the URL.  In future,
once the links attribute has been included in the message document, we can interrogate the content store to get the correct link.

# This is the 2nd commit message:

Stub external http requests in tests

# This is the 3rd commit message:

Conditionally include links hash in call to email-alert-api

Update the email alert model to include a links hash containing a content
ID, if one is present in the payload received from a client app.

This enables support for email alert notifications triggered by either
tags or links. This is an intermediate state as we migrate to using
links/content IDs only.

# This is the 2nd commit message:

Tidy up