Use the message queue gem
This commit makes email-alert-service use the
`govuk_message_queue_consumer` gem. There should be no change in
functionality.

Integration testing of the entire process have proven to be difficult
and brittle. In the interest of progress, we have decided to skip that
for now. Instead of an end-to-end test, an extra check has been added
that verifies that we won't be sending out email alerts for anything
other than "major" updates.