Conditionally include links hash in call to email-alert-api
Update the email alert model to include a links hash containing a content
ID, if one is present in the payload received from a client app.

This enables support for email alert notifications triggered by either
tags or links. This is an intermediate state as we migrate to using
links/content IDs only.