Reverse changes introduced by https://github.com/alphagov/email-alert-service/pull/48
Not all content items have a parent element in the links hash.  This commit reverses all changes (except the changes to use the GOVUK_ENV environment variable in testing), while we re-address whether or not there is still an issue with email alerts pointing to the wrong URL in light of changes made to the publishing_api.