Check links hash for tags when determining whether or not to send email
This checks the incoming document's links hash for the presence of topic
or policy tags. It also cleans up some of the tests related to message
processing and the email alert api payload.

In message_processor_spec.rb in particular, this commit tries to make
the tests easier to follow by mutating the test data in each test case
(rather than setting up multiple, complete test documents at the top of
the spec). It preserves (and builds upon) the variety of scenarios the
processor is supposed to handle.