Allow documents with service manual topics to trigger emails
We're building email alerting into the service manual.

We need documents with a link to `service_manual_topics` to be sent to the email alert api so that they can be matched against possible subscriber lists.

I've also renamed the method from `contains_supported_tag?` to `contains_supported_attribute` because the method is used for checking both tags and links.