Send document_type to the email-alert-api
The email alert api allows for matching on document type, but the email alert service does not send the document type. The only way the document_type matcher can currently be used is by hitting it directly (as travel-advice-publisher does).

This changes the service to always send the document type to the email alert api. The document type should be irrelevant unless a subscriber list is set to match against it, so this should have no impact on existing subscriptions.