Send all service manual guides to email alert api
We want users to be able to subscribe to changes to the service manual standard points. Points are just guides where the service standard is the parent. We don't want to send all documents with parents to the email alert api. Instead we send all the guides and let the email alert api work it out.