Remove heartbeat stubs from specs where they're not needed
The specs for missing title and public_updated_at and the spec for a
non-english locale stubbed the message so it was a heartbeat meaning
the specs passed for the wrong reasons (heartbeat messages are always
ignored, regardless of other attributes).

Fixing these tests reveals that the processor will process messages with
no public_updated_at even though we had a test that suggested we didn't.
As public_updated_at is used in MessageIdentifier it would break if they
weren't present, so we add a new guard clause instead of removing the
test.