Avoid treating "null" values in title or public_updated_at as valid
We have seen documents that don't have a title where instead of the
attribute not being present in the JSON, it is present but has the value
"null".  When this is turned into a ruby hash we get a key with a nil
value.  Using Hash#fetch with a default value protects us a missing key
but not against a key with a nil value.  This meant that when asking
if a document had a title we'd treat the nil value as it having a title
and proceed to triggering the alert which would fail as we'd try to
use the nil title to build a MessageIdentifier.  This commit changes
the guard clauses to understand the difference between a missing key
and a nil value.