Test message processor against all valid content
This adds a test that runs the message processor against all random
content generated for all of our schemas. Its purpose is to find cases
where the processor can't handle messages from the queue.

For example, the processor assumes that `document_type` exists on the
incoming document[1]. If the publishing-api doesn't send this we would
see a nil-error in this app. But for it to be removed in the
publishing-api, changes would have to be made to the `notification`
schema[2]. The PR for that change would trigger our random tests,
exposing the nil-error.

The number 5 has no significance, other than that the random generator
is very slow and I didn't want to increase the test times that much.
With the current number of schemas this adds about 10 seconds to the
build time.

[1]:
https://github.com/alphagov/email-alert-service/blob/63b4cebff3714fb5824
2606d80661aaa9e470944/email_alert_service/models/message_processor.rb#L5
5
[2]: https://github.com/alphagov/govuk-content-schemas