Run tests against the latest content schemas on Jenkins
Check out the latest production version of the contents schemas before
running the unit tests on the CI server, to check that this service is
compatible with the latest schemas.

Also configure build parameters so that regular branch builds (including
master) are run against the production version of the content schemas,
but the build can also be used to test changes to the content schemas.