Ensure schema tests are run against the right version of the schemas
This build can be triggered by changes to the schema tests, so we need
to make sure that we check out the correct branch of the content schemas
to test against, rather than always testing against
`deployed-to-production`.

This fix adds a missing branch parameter when checking out the content
schemas.