Add taxon_tree to document links sent to email-alert-api
When a document has taxons present in its links hash, expand out the
ancestor tree of each taxon and include the unique set of content ids in
the payload sent to email-alert-api.

This will allow us to support email notifications for users who have
subscribed to a given taxon and its children.