Use first note from change_history in email body
We would like to include the change note in the email, if we can - it's
helpful for users to have the context of what's changed so they know if it's
relevant to them.

There are a variety of places in content items where change notes can be,
depending on the document type. We were previously only looking in one of
these places, which meant that most emails constructed by email-alert-service
didn't include a change note.

The publishing API tries to store change notes in a consistent format, and has
recently been fixed to ensure that that is done more often [1]. That means
that we should have a "change_history" array in the details hash in the
messages from the queue, in reverse chronological order. This commit switches
to using the first (most recent) change note from there in the email body.

[1]: https://github.com/alphagov/publishing-api/pull/834