Explicitly blacklist content from certain publishing apps
We're about to expand the range of content for which we trigger email alerts
through email-alert-service, but want to ensure that we don't accidentally
trigger duplicate emails for content from publishing apps which don't rely on
the message queue to send their emails. It seems reasonable to do this based
on publishing app rather than document type since it's based on the behaviour
of these two applications, and new document types can be added to specialist
publisher so it's cleaner to not have to maintain a list of them here.