Trigger emails for content based on email and government supertypes
These document supertypes have been added to support the migration of
Whitehall's email subscriptions from govuk-delivery to email-alert-api. These
fields are already passed on with the document if they're present, but now we
need to also pass on content which has these attributes regardless of whether
it has links or tags.

This migration is intended to get all of our email subscriptions modelled in
one way, in one database, and (almost) all based off content sent to the
publishing API. We're not intending to change behaviour for users at the
moment so are using these fields to maintain most of the existing behaviour.
This is still a long way from where we'd like email subscriptions to be, but
it's a step in the right direction which will enable further changes to be
made more easily in future.