Expand list of attributes to look for in links or tags
These attributes are used in the filtering for Whitehall's email
subscriptions, so we need to pass on any content which has these links.

Tagging to these attributes is done by content editors and they are mostly
only available in Whitehall, so we think we can use them fairly reliably to
identify content which should have emails sent. We're excluding organisations
from this list even though they're an important part of email subscriptions
because many kinds of content are tagged automatically to organisations, and
so letting all of them through would dramatically expand the scope of those
subscriptions, which we don't want to do without careful consideration. Some
examples of content for which Whitehall doesn't send emails but are tagged to
organisations:

- corporate info pages
- help pages
- contacts
- hmrc manual sections

There are some other document types for which Whitehall doesn't currently
trigger emails but probably should - including organisations here would let
them through but it's better to do that explicitly later on, rather than at
the same time as this migration:

- case studies
- detailed guides