Add check for base_path value being present
DocumentValidator checks that 3 required fields are present as keys in the
document, but doesn't check their values. MessageProcessor checks the values
of two of those three plus locale, but wasn't checking that base_path had a
useful value.

When we expanded the filtering to include more content to support the
Whitehall subscriptions, we started trying to process some messages for
Whitehall contacts, which are non-addressable content items. They have a nil
base_path (in the representation that comes through on the queue, anyway) so
were passing the earlier checks but erroring when trying to construct the
message body (in EmailAlertTemplate#make_url_from_document_base_path).

This commit expands the earlier checks to ensure that if a message has a nil
base_path it is acknowledged and we don't try to send an email for it.