Remove DocumentValidator
message.validate! was being called only after other, more extensive checks for
the same fields, which return early if those checks fail, and so this
validation is misleading rather than useful to us. We'd prefer to continue to
log these cases and acknowledge the messages rather than move the validate!
call above the checks to reject them - we think they'll be more visible to us
that way.

These fields are required for sending emails, but aren't always present in
messages for all document_types, for example redirects don't have a title and
non-addressable content items don't have a base_path. That means that the
message could be correctly formatted from the point of view of the content
item but not usable for email sending. We don't want to have to define what
makes messages for different document types valid in this app, so it seems
sensible to move to only checking for these fields in the context of deciding
whether or not to send an email.

A better longer term solution could be to use a different queue for sending
emails, which only has user-facing content which is relevant for emailing to
users (that might not include every page on the site), but that needs more
thought and time than we have right now.