Inline the code to generate the taxon tree
This brings the code to generate the taxon tree into this app. The
taxonomy gem was part of an effort to avoid duplication between
applications dealing with the taxonomy. Unfortunately we ran out of
time last quarter, which left us with an API that was still in flux
(https://github.com/alphagov/govuk_taxonomy_helpers/pull/11).

To avoid confusion, I think it's best to inline the code in the places
where we use the taxonomy helpers. We're going to be doing work around
the taxonomy in the next couple of months and I think it will be easier
to extract code than trying to nail the API right now.

https://trello.com/c/FVfbiKxn