Allow rabbitmq config to be passed in as env var
The rabbitmq hostnames are hardcoded in our deployment scripts, and this
means this application won't be able to work properly with rabbitmq when
running in Amazon (where we use different hostnames).

Allow the config file to be read in as ERB so we can set a conditional
based upon environment variables.

Update the configuration so we can replace the current deployment
scripts.