Require all models in environment
Prior to this commit models were required in a couple of places that as
luck would have it made them available where the were needed.

When trying to add another model (in another PR) it wasn't clear why it
was undefined in its specs but other models seemingly got required by
magic.

This commit loads all models in the environment which is included in
both spec_helper and bin/email_alert_service.