Requeue HTTP error responses from Email Alert API
This changes the behaviour of this service. Prior to this change if
EmailAlertAPI returned a HTTP error code (502, 504) this application
would terminate and then be restarted by upstart. These messages would
not be acknowledged so would be re-queued.

Instead this catches HTTP errors which aren't caught earlier in the
change and requeues the message on Rabbit while logging the error.

There is a risk here that we end up in an eternal loop if we keep
sending a request that cannot be resolved, but this PR does not
introduce this or make it worse (as far as I'm aware) as it was
already present in the previous behaviour of restarting the
application.