Update configuration file to allow multiple queues
We need to listen to multiple queues and so the original one-to-one
association between a connection and queue needs to be changed.

We can still assume one central exchange.

For each queue we have to supply the name, the routing key and the
the class we are going to use to process messages on the queue.