Fix exception handling
Currently when an exception occurs, we try and access a variable
called `e` which doesn't exist. This means the whole program crashes
but upstart doesn't restart it since it hasn't exited, it's just
blocking.

I think this is the reason why we've seen email-alert-service
sometimes stop processing the queue, and it seems related to when
RabbitMQ is unreachable in some way. Hopefully this fix should
mean that we can see what the error is next time.