Add rake task to create queues
This is to allow the creation of the queues that are part of the
rabbitmq config file. This is primarily used by end-to-end tests as a
means to set up the environment so tests can run.

Incidentally, it appears the queue configurations in the rabbitmq config
file are not used by anything besides this so could probably just be
hardcoded here and removed from that file. I wanted to minimise changes
though so I avoided that step.