Rescope email alerts for the service standard
Previously we added a rule that allowed for alerts to be triggered
when a service_manual_guide was published, in order to accommodate
the subscriptions to the service standard, which is a singleton index
page for several guides [1]. Allowing based on document type is not
ideal because it's unclear what subscriptions this might trigger
updates for. This replaces the original approach by fully articulating
the conditions for which an alert should trigger.

[1]: https://github.com/alphagov/email-alert-service/commit/1db3c0e34896eae8b7adfa938e11c509fd35e591