Format STDOUT logs as JSON
Previously we wrote logs in plaintext, which means they get replaced
with an error message when filebeat tries to process them as JSON, to
send to Kibana [1]. Rather than try to reconfigure filebeat for just
this one app, it makes more sense to be consistent: use JSON logs.

Note that we conventionally use "@timestamp" as the field name for
Rails logs [2], while "message" is a pre-existing field for other logs.
We also don't seem to report "severity" for things.

[1]: https://github.com/alphagov/govuk-puppet/blob/master/modules/filebeat/templates/prospector.yml.erb#L70
[2]: https://github.com/shadabahmed/logstasher#about-logstasher