[TOFIXUP] Add section on user needs
I figure it makes sense to outline the need we're building this feature
around and then situate unsubscriptions as a part of it.

It's true that no-one has asked for this yet, but we've not exposed any
users to this yet. It seems like a safe bet it's a bad user experience
to let someone subscribe to something specific, then disappear it when
it's unpublished without telling them.

The taxon's case was for a broader collection of guidance, so i can see
the need there was likely weaker!