Move shared validation checks up to message_processor
Previously we've only had one message processor, we are now about to add
a second. It would be sensible to re-use some of these validation checks
between them to avoid repitition.

I considered placing these in their own module as a helper, however this
repo leans heavily on an inheritance model, so it seemed to make more
sense to put things common to all MessageProcessors in the parent class.

Items specific to an individual processor can then be added as private
methods to a child class of MessageProcessor.

This does seem to clutter an otherwise fairly simple MessageProcessor
class however... I am open to dissenting views on the best approach
here?