Add uuidtools and method to generate v5_uuid
In rails apps, where we have access to Digest extended by activesupport
we can use a much easier method. However as this is a Plain Old Ruby
App, it's likely easier to pull in this tool, instead of trying to plug
in all of activesupport.

See here for comparison:
https://www.uuidgenerator.net/dev-corner/ruby

a v5 hash is required by the email-alert-api when we wish to send a
message. It will combine the content ID and publicly updated at into a
simple string we can use for comparison and prevent duplicate emails
being sent.