Rename LockHandler Parameters
Previously the LockHandler class has only been used for content change
emails, and there we were always passing it an email title parameter.

We want to re-use this class for unsubscription notifications which will
not have access to them same title but do have a uuid we can use.

As the string here is only used to render a name, we can use either.
Let's rename the parameter to something more permissive to avoid
confusion