Add SinglePageUnpublishingMessageProcessor
We will need a message processor, similar to the processor that watches
out for major changes.

This one will monitor for unpublication events:
- That have the required info to interact with the email ecosystem
(base_path, content_id and public_updated_at).
- That are in english (we are only providing english language updates*)
- That fall into one of three unpublishing scenarios we wish to provide
specific email alerts to users above.

We can then use this to remove all single page notification subscribers
by finding a subscriber_list_id via the content_id.

When clearing that list we can notify users to let them know why their
subscription has disappeared.

* # Why are we only providing english language updates?

This was a bit of a rabbit hole, so I'll document here what we found.

Single page notifications relies on content_ids to distingush content
and allow users to subscribe to specific pages. However content in
different locales (that renders on it's own page) shares a content ID
and is distinguished by a locale.

We tried to trace back the git history and ended up here:
https://github.com/alphagov/email-alert-service/pull/33

>  For now, we'll simply ignore any messages which
aren't in english.

We're unsure why this was, but can generate some theories.

It may be jarring to subscribe to a page in Welsh and then only recieve
English updates. There are some pages that do not have English content
(we mostly discovered world_news_articles that only have content in a
specific locale).

We also wondered if it was possible to unpublish an item in only one
locale, and if that might trigger notifications for pages that remain in
english.

In the end we decided that the single page notification feature would
only be applied to english locale pages anyway, meaning mirroring this
restriction here is unlikely to cause issues.

There is likely a need for localised content - however this is probably
best dealt with as part of a larger overhaul of how localised content
should work on GOV.UK, and the email notification system.

In particular perhaps by making SubscriberLists aware of the locale of
content and handling updates / unsubscription notifications for each
separately. That's out of scope for us now, but was interesting to
explore.