Fix email_alert_api bulk_unsubscribe takes params not hash
Also we released we need to pass along a govuk_request_id.

As part of debugging this, we've realised that when you run this in
local development (with govuk-docker) the govuk_request_id is not
populated with any dummy content.

On our servers, this gets populated by the CDN (we think). It doesn't
seem that there's a layer that injects a value here.

If you are trying to do local dev examining APIs calls between
email-alert-service and the email-alert-api for the bulk unsubscribe
endpoint. You will get UnprocessibleEntity errors returned. These are
not the UnprocessibleEntitys that we are handling there in the
controller but are bubbling up from the Message validation that requires
a govuk_request_id.

We should probably stort out a way of mocking in govuk_request_ids for
local dev setups at some point in the future.