Correctly parse public_updated_at timestamps
Because we were going via `Time.now` instead of `Time.iso8601`, we
were losing information and ending up with midnight on the 1st of
January (this year).

I've updated the tests to go more directly from a `Time` object to the
formatted time, which catches this, and then fixed the code.