Add UpdateSubscriberList class
WHAT
====

The GOV.UK Accounts team added a feature where logged in users could
subscribe to email notifications for individual pages.

SubscriberLists store the titles of those pages so that we can:
- Render a list of subscriptions in the user's manage dashboard
- Include the title of the subscriber list in emails alerting the user
that the page has been unpublished. (Especially as there's no-longer a
content item we can query for title at that point).

We record these titles when the SubscriberList is first made, however
there has not been a mechanism to update it. This is a problem as
content titles do change over time.

This PR adds a class we can use to gather information from major and
minor change messages and update the subscriber list titles using a new
[Email Alert API endpoint](1)

Other parts of email-alert-service make calls to services that use GDS
API adaptors. However they seem not to use the test helpers that ship
with the gem through webmock.

A note on test strategy
======================

In previous work we decided to keep the testing strategy consistent
across the repo, and also avoided gds-api-test helpers. Instead we
tested with double and spys.

This got us into a bit of trouble when our parameters and payloads were
not matching the gds-api-adaptors, and reminded us how helpful those
test helpers were.

So we've introduced stubs and webmock for this test, aware this
introduces an inconsistent test strategy.

We intend to go back and refactor tests for the recently introduced
unpublication alert work. And hopefully refactor the rest of the test
suite to.

If you are far in the future and that didn't happen. Sorry! But
hopefully you at least know why these tests don't look like the others.

(1): https://github.com/alphagov/email-alert-api/pull/1704

Co-authored-by: Alex Whitehead-Smith <alex.whitehead-smith@digital.cabinet-office.gov.uk>
Co-authored-by: Huw Diprose <huw.diprose@digital.cabinet-office.gov.uk>