Add UpdateSubscriberList class
WHAT
====

The GOV.UK Accounts team added a feature where logged in users could
subscribe to email notifications for individual pages.

SubscriberLists store the titles of those pages so that we can:
- Render a list of subscriptions in the user's manage dashboard
- Include the title of the subscriber list in emails alerting the user
that the page has been unpublished. (Especially as there's no-longer a
content item we can query for title at that point).

We record these titles when the SubscriberList is first made, however
there has not been a mechanism to update it. This is a problem as
content titles do change over time.

This PR adds a class we can use to gather information from major and
minor change messages and update the subscriber list titles using a new
[Email Alert API endpoint](1)

Other parts of email-alert-service make calls to services that use GDS
API adaptors. However they seem not to use the test helpers that ship
with the gem through webmock.

(1): https://github.com/alphagov/email-alert-api/pull/1704

Co-authored-by: Alex Whitehead-Smith <alex.whitehead-smith@digital.cabinet-office.gov.uk>
Co-authored-by: Huw Diprose <huw.diprose@digital.cabinet-office.gov.uk>