Add EmailAlertService.services(:redis) to unpublishing_consumer
Looking back through past commits there is a really useful discussion in
PR #470.

> As part of the work to enable continuous deployment of all GOV.UK
apps we need to add a check for connectivity to Redis. Since Email
Alert Service is not a web app, this check had to be added to the rake
task that is called at startup time. This commit adds a call to the
Redis service that will throw an error if it does not exist.

We should have added this here.

More info:
=========
https://github.com/alphagov/email-alert-service/pull/470