Update subscriber list description on content change
Extend the processors added in #493 to also check if the subscriber list
description has changed with a major or minor content change. If it has,
send the new version to email alert API.

These descriptions will be included in the emails sent when a page is
unpublished, so it's important that they're kept up to date.

Depends on https://github.com/alphagov/email-alert-api/pull/1709