Only show page summary if description is present
All new single page subscriber lists have a description when they're
created, and we've backfilled descriptions for ones that were created
prior to this. However, there are no requirements on the email alert
API side for a single page list to have a description because we want to
keep the API free from any special cases.

The description comes from the content item, and content items for
single pages should all have a description which is enforced by GOV.UK's
content schemas when it hits Publishing API and the content store.

There is still a small chance that a list gets created without a
description, or that the description is removed from the list. In these
cases we should gracefully handle the missing description by not
rendering that section of the email body, but also raising an error in
Sentry so someone can investigate.